+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 > All notable changes to this project will be documented in this file.
 
+## [0.3.2] - 2022-05-29
+
+- Add unit tests & run them via Github Actions
+- Add MIT license
+- Move the menu button to the left
+- Add auto theme what listens to your OS settings
+
 ## [0.3.1] - 2022-05-03
 
 - Display nice to have features (theme, language)